requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-docx==1.1.2
python-multipart==0.0.20
pydantic-settings==2.7.1
python-dotenv==1.0.1
jinja2==3.1.5
aiofiles==24.1.0
email-validator==2.2.0
